requirements.txt
# AI Development Requirements

# Core AI/ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0

# OpenAI API
openai>=1.0.0
google-generativeai>=0.3.0

# Data Processing
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: Web framework for API
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0


